Fixed float to int deprecation notice

+++ ProcessWorker.php
@@ -72,10 +72,7 @@ Response: {
     "type": "endInput",
     "pid": 12345
 }
-
-
 */
-
 
 class ProcessWorker {
     protected $process_handles = [];        // [pid][control|stdin|stdout|stderr] = resource
@@ -155,7 +152,7 @@ class ProcessWorker {
             if(!$this->active) {
                 $diff = microtime(true) - $time;
                 if($diff < 0.1) {
-                    usleep(100000 - ($diff * 1000000));
+                    usleep((int)(100000 - ($diff * 1000000)));
                     break;
                 }
             }
